# backend/docs/swagger.yaml
basePath: /
definitions:
...
  CreateWalletRes:
    properties:
      data:
        $ref: '#/definitions/domain.Wallet'
    required:
    - data
    type: object
...
  DeleteWalletRes:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  GetWalletRes:
    properties:
      data:
        $ref: '#/definitions/domain.Wallet'
    required:
    - data
    type: object
...
  UpdateWalletRes:
    properties:
      data:
        $ref: '#/definitions/domain.Wallet'
    required:
    - data
    type: object
...
  domain.Wallet:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - created_at
    - id
    - name
    - type
    - updated_at
    - user_id
    type: object
...
  dto.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    required:
    - code
    - message
    type: object
...
  dto.Wallet:
    properties:
      data:
        $ref: '#/definitions/domain.Wallet'
    required:
    - data
    type: object
...
  wallet.CreateWalletReq:
    properties:
      name:
        type: string
      type:
        type: string
    required:
    - name
    - type
    type: object
  wallet.ListWalletsRes:
    properties:
      wallets:
        items:
          $ref: '#/definitions/dto.Wallet'
        type: array
    required:
    - wallets
    type: object
  wallet.UpdateWalletReq:
    properties:
      name:
        type: string
      type:
        type: string
    required:
    - name
    - type
    type: object
host: localhost:8080
info:
  contact: {}
  description: API untuk aplikasi manajemen keuangan dan pencatatan emas - ARTA
  title: ARTA API
  version: "1.0"
paths:
...
  /api/wallet:
    get:
      consumes:
      - application/json
      operationId: ListWallets
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.ListWalletsRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      security:
      - Bearer: []
      tags:
      - wallet
    post:
      consumes:
      - application/json
      operationId: CreateWallet
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Unique key per submission attempt for safe retry (UUID recommended)
        in: header
        name: Idempotency-Key
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/wallet.CreateWalletReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CreateWalletRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      security:
      - Bearer: []
      tags:
      - wallet
  /api/wallet/{id}:
    delete:
      consumes:
      - application/json
      operationId: DeleteWallet
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DeleteWalletRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      security:
      - Bearer: []
      tags:
      - wallet
    get:
      consumes:
      - application/json
      operationId: GetWallet
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetWalletRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      security:
      - Bearer: []
      tags:
      - wallet
    put:
      consumes:
      - application/json
      operationId: UpdateWallet
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/wallet.UpdateWalletReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UpdateWalletRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      security:
      - Bearer: []
      tags:
      - wallet
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN''"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"